requirements.txt
fastapi==0.95.2
jinja2==3.1.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
starlette==0.27.0
pydantic==1.10.4